requirements.txt
absl-py==0.5.0
amqp==2.3.2
argh==0.26.2
args==0.1.0
asn1crypto==0.24.0
astor==0.7.1
atari-py==0.2.6
atomicwrites==1.2.1
attrs==19.3.0
backcall==0.1.0
backports.weakref==1.0rc1
basemap==1.1.0
beautifulsoup4==4.6.3
billiard==3.5.0.5
bleach==3.1.0
blis==0.2.4
boto==2.49.0
boto3==1.9.4
botocore==1.12.4
bottle==0.12.13
box2d-py==2.3.8
branca==0.3.0
bz2file==0.98
celery==4.2.1
certifi==2018.4.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
clint==0.5.1
cloudpickle==0.5.6
colorama==0.4.1
configparser==4.0.2
convertdate==2.1.3
CProfileV==1.0.7
cryptography==2.7
cssselect==1.0.3
cycler==0.10.0
cymem==2.0.2
Cython==0.28.5
dask==0.19.2
decorator==4.4.1
defusedxml==0.6.0
docker-pycreds==0.4.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
ephem==3.7.7.0
feedparser==5.2.1
ffmpeg==1.4
Flask==1.0.2
floyd-cli==0.11.7
folium==0.5.0
future==0.17.1
gast==0.2.0
gensim==3.5.0
gitdb2==2.0.6
GitPython==3.0.4
googlemaps==2.5.1
gql==0.1.0
graphql-core==2.2.1
graphviz==0.10.1
grpcio==1.15.0
gym==0.10.8
h5py==2.8.0
holidays==0.9.11
html5lib==0.9999999
hyperopt==0.1.1
idna==2.7
imageio==2.4.1
importlib-metadata==0.23
inflection==0.3.1
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.3
JPype1==0.6.3
JSAnimation==0.1
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
kaggle==1.5.3
Keras==2.0.6
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
keras-rl==0.2.0
kombu==4.2.2
konlpy==0.5.1
lightgbm==2.2.3
lunardate==0.2.0
lxml==4.2.5
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==2.15.4
matplotlib==2.0.0
mecab-python===0.996-ko-0.9.2-msvc
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==7.2.0
moviepy==0.2.3.5
mpmath==1.0.0
murmurhash==1.0.1
nbconvert==5.6.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.2
nltk==3.2.3
nose==1.3.7
notebook==6.0.2
numpy==1.15.2
nvidia-ml-py3==7.352.0
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
pandas==0.25.1
pandas-datareader==0.7.4
pandocfilters==1.4.2
parso==0.5.1
pathlib2==2.3.2
pathtools==0.1.2
patsy==0.5.0
pickleshare==0.7.5
Pillow==5.2.0
plac==0.9.6
pluggy==0.7.1
preshed==2.0.1
prometheus-client==0.7.1
promise==2.2.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psutil==5.6.5
py==1.7.0
pyasn1==0.4.7
pycparser==2.19
pydot==1.2.3
pydot-ng==1.0.0
pyglet==1.2.4
Pygments==2.4.2
pymongo==3.7.1
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.3.1
pyproj==1.9.5.1
pyrsistent==0.15.5
pyshp==1.2.12
pysolar==0.8
pystan==2.19.0.0
pytest==3.8.2
python-dateutil==2.8.1
python-slugify==3.0.0
pytz==2018.7
PyWavelets==1.0.0
pywin32==227
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.1.1
qtconsole==4.5.5
Quandl==3.4.8
raven==6.9.0
redis==3.0.1
requests==2.19.1
requests-toolbelt==0.8.0
Rx==1.6.1
s3transfer==0.1.13
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.7.1
selenium==3.0.0
Send2Trash==1.5.0
sentry-sdk==0.13.2
setuptools-git==1.2
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.13.0
sklearn-extensions==0.0.2
smart-open==1.6.0
smmap2==2.0.5
spacy==2.1.8
srsly==0.1.0
statsmodels==0.9.0
subprocess32==3.5.4
sympy==1.0
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
textblob==0.15.1
theano==1.0.2
thinc==7.0.8
toolz==0.9.0
torch==1.2.0+cpu
torchvision==0.4.0+cpu
tornado==6.0.3
tqdm==4.28.1
traitlets==4.3.3
urllib3==1.23
vine==1.1.4
wandb==0.8.15
wasabi==0.2.2
watchdog==0.9.0
watermark==1.6.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
xgboost==0.80
xlrd==1.2.0
XlsxWriter==1.1.2
zipp==0.6.0
